+++ examples/fsm-examples/src/lib/simple.ts
@@ -20,9 +20,13 @@ const simpleFsm: FiniteStateMachine<number, string, string> = {
 		const invocations = state > 10 ? [] : [{ actionId: 'triple', argument: state }]
 		return { state, invocations, output: state.toFixed(2) }
 	},
-	async onError(err) {
-		console.log('ERROR!')
-		console.log(err)
+	onExecutionError(state, input, error) {
+		console.log('ERROR while executing state machine:')
+		console.log({ state, input, error })
+	},
+	onInvocationError(actionId, argument, error) {
+		console.log('ERROR while executing state machine:')
+		console.log({ actionId, argument, error })
 	},
 }